removed modifications to yawning_titan and replace broken pyinstaller with custom release on rumbelows repo
