+++ yawning_titan_gui/management/commands/run_gui.py
@@ -5,7 +5,7 @@ class Command(BaseCommand):
 
     def handle(self, *args, **kwargs):
         from flaskwebgui import FlaskUI
-        from yawning_titan_gui.wsgi import application as app    
+        from yawning_titan_server.wsgi import application as app    
 
         print(f"running app with {sys.executable}")
         FlaskUI(app=app, server="django").run()
