+++ yawning_titan_gui/management/commands/setup.py
@@ -7,7 +7,7 @@ class Command(BaseCommand):
         print("Running setup...")
         from yawning_titan import DATA_DIR
         from yawning_titan_gui import _YT_GUI_ROOT_DIR        
-        from dirs import _create_app_dirs,_copy_package_data_notebooks_to_notebooks_dir
+        from setup import _create_app_dirs,_copy_package_data_notebooks_to_notebooks_dir
 
         _create_app_dirs()
         _copy_package_data_notebooks_to_notebooks_dir()
